Renamed files with + in them to be alphanumeric and baselined to roughly r2 or http://unittestpp.googlecode.com/svn/trunk/

+++ src/TestRunner.cpp
@@ -34,7 +34,7 @@ TestRunner::~TestRunner()
 
 int TestRunner::Finish() const
 {
-    float const secondsElapsed = m_timer->GetTimeInMs() / 1000.0f;
+    float const secondsElapsed = static_cast<float>(m_timer->GetTimeInMs() / 1000.0);
     m_reporter->ReportSummary(m_result->GetTotalTestCount(), 
 							  m_result->GetFailedTestCount(), 
 							  m_result->GetFailureCount(), 
@@ -60,7 +60,7 @@ void TestRunner::RunTest(TestResults* const result, Test* const curTest, int con
 
 	curTest->Run();
 
-	int const testTimeInMs = testTimer.GetTimeInMs();
+	double const testTimeInMs = testTimer.GetTimeInMs();
 	if (maxTestTimeInMs > 0 && testTimeInMs > maxTestTimeInMs && !curTest->m_timeConstraintExempt)
 	{
 	    MemoryOutStream stream;
@@ -70,7 +70,7 @@ void TestRunner::RunTest(TestResults* const result, Test* const curTest, int con
 	    result->OnTestFailure(curTest->m_details, stream.GetText());
 	}
 
-	result->OnTestFinish(curTest->m_details, testTimeInMs/1000.0f);
+	result->OnTestFinish(curTest->m_details, static_cast<float>(testTimeInMs/1000.0));
 }
 
 }
